fix: health dockerhub call
