+++ packages/backend/src/services/HealthService/HealthService.ts
@@ -1,20 +1,11 @@
 import { HealthState, LightdashInstallType, LightdashMode } from 'common';
-import fetch from 'node-fetch';
+import { getDockerHubVersion } from '../../clients/DockerHub/DockerHub';
 import { LightdashConfig } from '../../config/parseConfig';
 import { getMigrationStatus } from '../../database/database';
 import { UnexpectedDatabaseError } from '../../errors';
 import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
 import { UserModel } from '../../models/UserModel';
 import { VERSION } from '../../version';
-
-const filterByName = (result: { name: string }): boolean =>
-    /[0-9.]+$/.test(result.name);
-
-const sorterByDate = (
-    a: { last_updated: string },
-    b: { last_updated: string },
-): number =>
-    Number(new Date(b.last_updated)) - Number(new Date(a.last_updated));
 
 type HealthServiceDependencies = {
     userModel: UserModel;
@@ -49,19 +40,6 @@ export class HealthService {
             );
         }
 
-        let latestVersion: string | undefined;
-        try {
-            const response = await fetch(
-                'https://hub.docker.com/v2/repositories/lightdash/lightdash/tags',
-                { method: 'GET' },
-            );
-            latestVersion = (await response.json()).results
-                .filter(filterByName)
-                .sort(sorterByDate)[0].name;
-        } catch {
-            latestVersion = undefined;
-        }
-
         const needsProject = !(await this.projectModel.hasProjects());
 
         const localDbtEnabled =
@@ -77,7 +55,7 @@ export class HealthService {
             localDbtEnabled,
             defaultProject: undefined,
             isAuthenticated,
-            latest: { version: latestVersion },
+            latest: { version: getDockerHubVersion() },
             rudder: this.lightdashConfig.rudder,
             sentry: this.lightdashConfig.sentry,
             intercom: this.lightdashConfig.intercom,
